refactor: extract client.rs
